requirements.txt
# Core ML Libraries
torch>=2.0.0
# Note: Use compatible versions to avoid dispatch_batches error
# If you encounter Accelerator errors, try: transformers==4.38.2 accelerate==0.28.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0,<0.29.0  # Constrain to avoid dispatch_batches compatibility issues
evaluate>=0.4.0
tokenizers>=0.15.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Experiment Tracking (Optional)
wandb>=0.15.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=6.5.0

# Utilities
tqdm>=4.65.0

